+++ Pandora/Paradigm/Structure/Stack.hs
@@ -1,5 +1,6 @@
-module Pandora.Paradigm.Structure.Stack (Stack, push, top, pop, filter, linearize) where
+module Pandora.Paradigm.Structure.Stack (Stack, push, top, pop) where -- , filter, linearize
 
+import Pandora.Core.Functor (type (:.:))
 import Pandora.Core.Morphism ((.), ($))
 import Pandora.Paradigm.Basis.Twister (Twister ((:<)), unwrap)
 import Pandora.Paradigm.Basis.Maybe (Maybe (Just, Nothing))
@@ -15,21 +16,13 @@ import Pandora.Pattern.Functor.Bindable (Bindable ((>>=)))
 import Pandora.Pattern.Object.Setoid (bool)
 
 -- | Linear data structure that serves as a collection of elements
-type Stack = (Twister :>: Maybe)
+type Stack a = (Maybe :.: Twister Maybe) a
 
 push :: a -> Stack a -> Stack a
-push x (Y struct) = (Y $ (:<) x . Just <$> struct) <+> point x
+push x stack = ((:<) x . Just <$> stack) <+> (point . point) x
 
 top :: Stack a -> Maybe a
-top (Y struct) = extract <$> struct
+top stack = extract <$> stack
 
 pop :: Stack a -> Stack a
-pop (Y struct) = Y $ struct >>= unwrap
-
-filter :: Predicate a -> Stack a -> Stack a
-filter (Predicate p) = Y . fold Nothing
-	(\x s -> bool s (Just $ x :< s) $ p x)
-
--- | Transform any traversable structure into a stack
-linearize :: Traversable t => t a -> Stack a
-linearize = Y . fold Nothing (\x -> Just . (:<) x)
+pop stack = stack >>= unwrap