+++ CHANGELOG.md
@@ -62,3 +62,4 @@
 
 # 0.1.7
 * Define `ifelse` function for `Boolean` values
+* Change `Stack` and `Nonempty` definitions, temporarily remove `filter` and `linearize`